Add try/catch for E211
In the case that nnn is used to remove a file while that file is in an
active buffer, Vim will raise E211 with trace on reentry from nnn. The
trace is somewhat of an eye sore but the E211 message is useful. This
PR adds a try catch for E211, removes the trace and reraises E211.

Fixes: #41